💄design: 공통 컴포넌트 스타일 컴포넌트 컨벤션 적용
