+++ grass-diary/src/components/Comment/CommentDisplay.tsx
@@ -1,6 +1,4 @@
-import styled from 'styled-components';
-import { semantic } from '@styles/semantic';
-import { TYPO } from '@styles/typo';
+import * as S from '@styles/component/Comment/CommentDisplay.style';
 import { useEffect, useRef, useState } from 'react';
 import { useWriterProfile } from '@hooks/api/useWriterProfile';
 import { useCommentActions } from '@state/comment/CommentStore';
@@ -38,110 +36,41 @@ const CommentDisplay = ({ comment, parentId }: CommentDisplayProps) => {
   }, [date, comment.createdDate]);
 
   return comment.deleted ? (
-    <CommentWrap $isMe={memberId === comment.memberId}>
-      <WriterWrap>
+    <S.CommentItem $isMe={memberId === comment.memberId}>
+      <S.WriterBox>
         {comment.depth ? <ReplyIcon /> : null}
-        <DeletedContent>{COMMENT.deleted}</DeletedContent>
-      </WriterWrap>
-    </CommentWrap>
+        <S.DeletedText>{COMMENT.deleted}</S.DeletedText>
+      </S.WriterBox>
+    </S.CommentItem>
   ) : (
-    <CommentWrap onClick={reply} $isMe={memberId === comment.memberId}>
-      <CommentTop>
-        <WriterWrap>
+    <S.CommentItem onClick={reply} $isMe={memberId === comment.memberId}>
+      <S.TopBox>
+        <S.WriterBox>
           {comment.depth ? <ReplyIcon /> : null}
-          <WriterProfile src={writer?.profileImageURL} />
-          <WriterName $isMe={memberId === comment.memberId}>
+          <S.ProfileImage src={writer?.profileImageURL} />
+          <S.NameText $isMe={memberId === comment.memberId}>
             {writer?.nickname}
-          </WriterName>
-          <CreateTime>
+          </S.NameText>
+          <S.TimeText>
             {isToday
               ? comment.createdAt
               : comment.createdDate
               ? `20${comment.createdDate}`
               : ''}
-          </CreateTime>
-        </WriterWrap>
+          </S.TimeText>
+        </S.WriterBox>
         <div ref={setting}>
           <CommentSetting
             commentId={comment.commentId}
             writerId={comment.memberId}
           />
         </div>
-      </CommentTop>
-      <CommentContent $isReply={comment.depth ? true : false}>
+      </S.TopBox>
+      <S.ContentBox $isReply={comment.depth ? true : false}>
         {comment.content}
-      </CommentContent>
-    </CommentWrap>
+      </S.ContentBox>
+    </S.CommentItem>
   );
 };
 
 export default CommentDisplay;
-
-const CommentWrap = styled.div<{ $isMe: boolean }>`
-  display: flex;
-  padding: var(--gap-sm, 0.75rem) var(--gap-md, 1rem);
-  flex-direction: column;
-  align-items: flex-start;
-  gap: var(--gap-sm, 0.75rem);
-  align-self: stretch;
-
-  border-radius: var(--radius-sm, 0.75rem);
-  border: ${props =>
-    props.$isMe
-      ? 'none'
-      : `var(--stroke-thin, 1px) solid
-    ${semantic.light.border.transparent.assistive}`};
-
-  background: ${props =>
-    props.$isMe ? semantic.light.fill.transparent.assistive : 'none'};
-`;
-
-const CommentTop = styled.div`
-  display: flex;
-  align-items: center;
-  gap: var(--gap-md, 1rem);
-  align-self: stretch;
-`;
-
-const WriterWrap = styled.div`
-  display: flex;
-  align-items: center;
-  gap: var(--gap-2xs, 0.5rem);
-  flex: 1 0 0;
-`;
-
-const WriterProfile = styled.img`
-  width: 1.5rem;
-  height: 1.5rem;
-  flex-shrink: 0;
-  border-radius: var(--radius-empty, 1.5rem);
-  object-fit: cover;
-  background: ${semantic.light.fill.transparent.alternative};
-`;
-
-const WriterName = styled.p<{ $isMe: boolean }>`
-  ${TYPO.label2}
-  color: ${props =>
-    props.$isMe
-      ? semantic.light.accent.solid.hero
-      : semantic.light.object.solid.normal};
-`;
-
-const CreateTime = styled.p`
-  ${TYPO.caption1}
-  color: ${semantic.light.object.transparent.assistive};
-`;
-
-const CommentContent = styled.div<{
-  $isReply?: boolean;
-}>`
-  ${TYPO.body1}
-  ${props => props.$isReply && `padding-left: var(--gap-2xl, 2rem);`}
-  color: ${semantic.light.object.solid.normal};
-  white-space: pre;
-`;
-
-const DeletedContent = styled.p`
-  ${TYPO.body1}
-  color: ${semantic.light.object.transparent.alternative};
-`;
