+++ grass-diary/src/components/Layout/Layout.tsx
@@ -1,8 +1,7 @@
 import { Outlet } from 'react-router-dom';
 import Header from './Header';
 import Footer from './Footer';
-import Toast from './Toast';
-import Modal from '@components/modal/Modal';
+import { Toast, Modal } from '@components/index';
 
 const Layout = () => {
   return (
