+++ grass-diary/src/components/Notification/Callout.tsx
@@ -1,0 +1,17 @@
+import * as S from '@styles/component/Notification/Callout.style';
+import { ReactComponent as Info } from '@svg/info.svg';
+
+type CalloutProps = {
+  message: string;
+};
+
+const Callout = ({ message }: CalloutProps) => {
+  return (
+    <S.Container>
+      <Info />
+      <S.Text>{message}</S.Text>
+    </S.Container>
+  );
+};
+
+export default Callout;
