+++ grass-diary/src/components/index.ts
@@ -14,6 +14,7 @@ export { default as Layout } from './Layout/Layout';
 export { default as Divider } from './Divider';
 export { default as Menus } from './Button/Menus';
 export { default as Menu } from './Button/Menu';
-export { default as Modal } from './modal/Modal';
-export { default as CustomButton } from './modal/CustomButton';
-export { default as Callout } from './callout/Callout';
+export { default as Modal } from './Notification/Modal';
+export { default as CustomButton } from './Notification/CustomButton';
+export { default as Callout } from './Notification/Callout';
+export { default as Toast } from './Notification/Toast';
