+++ grass-diary/src/styles/component/Comment/CommentSetting.style.tsx
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+
+export const SettingBox = styled.div`
+  padding: var(--gap-4xs, 0.25rem);
+`;
