+++ grass-diary/src/styles/component/Comment/Comments.style.tsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const CommentList = styled.ul`
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: var(--gap-xs, 0.625rem);
+  align-self: stretch;
+`;
