+++ grass-diary/src/styles/component/Like/NormalLike.style.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export const Box = styled.div<{ $justifyContent: string }>`
+  display: flex;
+  gap: 0.625rem
+  justify-content: ${props => props.$justifyContent},
+ 
+`;
